% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxm_test.R
\name{boxm.test}
\alias{boxm.test}
\title{Box-M test statistic}
\usage{
boxm.test(x, y, type = "chi.squared")
}
\arguments{
\item{x}{data matrix of first group with rows representing the samples and columns representing variables.}

\item{y}{data matrix of second group with same number of columns as \code{x}}

\item{type}{Distribution to use for calculating the p-value. Possible values are "chi.squared"(Default) and "F"}
}
\value{
A list with two values
\describe{
\item{test.statistic}{The test statistic value}
\item{p.value}{The p-value computed using the distribution specified in \code{type}}
}
}
\description{
This function computes the Box M test statistic for testing equality of covariance matrices. The p-value is computed using either a chi-squared or F approximation.
}
\examples{
library(mvtnorm)
x = rmvnorm(n = 100, mean = numeric(10), sigma = diag(10))
y = rmvnorm(n = 100, mean = numeric(10), sigma = diag(10))
boxm.test(x, y)
}
\references{
T. W. Anderson. An introduction to multivariate statistical analysis. Wiley Series in Probability and Statistics, 3rd edition, 2003.
}
