% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cai_statistic.R
\name{cai.statistic}
\alias{cai.statistic}
\title{Cai test statistic}
\usage{
cai.statistic(x, y, alpha = 0.05)
}
\arguments{
\item{x}{data matrix of first group with rows representing the samples and columns representing variables.}

\item{y}{data matrix of second group with same number of columns as \code{x}}

\item{alpha}{Significance level of the test (default value is 0.05)}
}
\value{
A list with two values
\describe{
\item{test.statistic}{The test statistic value}
\item{decision}{A binary response of whether the null hypothesis is rejected or not rejected}
}
}
\description{
This function computes the test statistic for testing equality of covariance matrices as described in the Cai et al. paper (need citation)
}
\examples{
library(mvtnorm)
x = rmvnorm(n = 20, mean = numeric(100), sigma = diag(100))
y = rmvnorm(n = 10, mean = numeric(100), sigma = diag(100))
cai.statistic(x, y)
}
\references{
T. Cai, W. Liu, and Y. Xia. Two-sample covariance matrix testing and support recovery in high-dimensional and sparse settings. Journal of the American Statistical Association, 108(501):265–277, 2013.
}
