% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramp-package.R
\docType{package}
\name{cramp-package}
\alias{cramp-package}
\title{CRAMP: \strong{C}ovariance matrix testing using \strong{RA}ndom \strong{M}atrix \strong{P}rojections}
\description{
\packageDESCRIPTION{cramp}
}
\details{
The CRAMP package provides functions to perform hypothesis tests involving the covariance matrix of multivariate distributions. For the one-sample case, tests for uniformity and sphericity are provided. In the two sample case, the test is for equality of two covariance matrices. Functions to compute several traditional Gaussian likelihood-based and high-dimensional test statistics are provided in this package.
}
\section{Functions}{

\packageIndices{cramp}
}

\references{
{Need to add arxiv submission}
}
