% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramp_statistic.R
\name{cramp.statistic}
\alias{cramp.statistic}
\title{CRAMP test statistic}
\usage{
cramp.statistic(
  x,
  y = NULL,
  nrep = 1000,
  m = 1,
  parallel = TRUE,
  ncores = NULL
)
}
\arguments{
\item{x}{data matrix of first group with rows representing samples and columns representing variables.}

\item{y}{data matrix of second group with same number of columns as \code{x}. If not provided (default value is NULL), then the one-sample tests are performed as specified.}

\item{nrep}{number of repetitions for each average of pvalues}

\item{m}{Dimension to randomly project the data (Default value is 1)}

\item{parallel}{Logical variable - whether to parallelize or not (default is \code{TRUE})}

\item{ncores}{Number of cores to use (If NULL and parallel is TRUE, use 75\% of all cores). If \code{parallel = FALSE}, this argument will be ignored.}
}
\value{
The average p-value based on \code{nrep} bootstrap random projections.
}
\description{
This function computes the test statistic based on random projections and the Hotelling's \eqn{T^2} test statistic.
}
\examples{
\dontrun{
     x = rmvnorm(n = 20, mean = numeric(1000), sigma = diag(runif(1000)))
     y = rmvnorm(n = 20, mean = numeric(1000), sigma = diag(runif(1000)))
     cramp.statistic(x, y, m = 5)
}
}
