% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev_critvalue.R
\name{gev.critvalue}
\alias{gev.critvalue}
\title{Wu-Li critical value}
\usage{
gev.critvalue(T.n, nrand, sig.level)
}
\arguments{
\item{T.n, nrand, sig.level}{Intermediary quantities computed in the \code{\link{wu.li.statistic}}}
}
\value{
The critical value for rejecting/accepting the null hypothesis as defined in Wu and Li (2020)
}
\description{
Function to compute critical values for the \code{\link{wu.li.statistic}}
}
\references{
T.-L. Wu and P. Li. Projected tests for high-dimensional covariance matrices. Journal of Statistical Planning and Inference, 207:73 – 85, 2020. ISSN 0378-3758.
}
\seealso{
\code{\link{wu.li.statistic}}
}
