% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/li_chen_test.R
\name{li.chen.statistic}
\alias{li.chen.statistic}
\title{Li-Chen test statistic}
\usage{
li.chen.statistic(x, y)
}
\arguments{
\item{x}{data matrix with rows representing samples and columns representing variables}

\item{y}{data matrix of second group with same number of columns as \code{x}}
}
\value{
A list with two values
\describe{
\item{test.statistic}{The test statistic value}
\item{p.value}{The p-value}
}
}
\description{
This function computes the test statistic for testing equality of covariance matrices as described by Li and Chen (need citation)
}
\examples{
library(mvtnorm)
x = rmvnorm(n = 20, mean = numeric(100), sigma = diag(100))
y = rmvnorm(n = 20, mean = numeric(100), sigma = diag(100))
li.chen.statistic(x, y)
}
\references{
J. Li and S. X. Chen. Two sample tests for high-dimensional covariance matrices. Annals of Statistics, 40(2):908–940, 2012.
}
