% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_tests.R
\name{lrt.statistic}
\alias{lrt.statistic}
\title{One-sample likelihood ratio tests}
\usage{
lrt.statistic(x, test = "identity")
}
\arguments{
\item{x}{data matrix with rows representing samples and columns representing variables}

\item{test}{The type of test being performed - "identity"(default) or "sphericity".}
}
\value{
A list with two values
\describe{
\item{test.statistic}{The test statistic value}
\item{p.value}{The p-value}
}
}
\description{
The traditional likelihood ratio test statistic for the one-sample hypothesis of covariance matrix.
}
\examples{
library(mvtnorm)
# Test for identity
x = rmvnorm(n = 20, mean = numeric(100), sigma = diag(100))
lrt.statistic(x, test = "identity")
# Test for sphericity
y = rmvnorm(n = 10, mean = runif(100, 1, 3), sigma = diag(rgamma(100, 10, 3)))
lrt.statistic(y, test = "sphericity")
}
\references{
T. W. Anderson. An introduction to multivariate statistical analysis. Wiley Series in Probability and Statistics, 3rd edition, 2003.
}
