% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw_statistics.R
\name{lw.statistic}
\alias{lw.statistic}
\title{Ledoit-Wolf test statistic}
\usage{
lw.statistic(x, test = "identity")
}
\arguments{
\item{x}{data matrix with rows representing samples and columns representing variables}

\item{test}{The type of test being performed - "identity"(default) or "sphericity". The test statistics are modified versions of the \code{\link{john.nagao.statistic}}. If \code{test == "identity"}, the test statistic is equal to the Nagao test statistic and when \code{test == "sphericity"}, the Ledoit-Wolf test statistic computed is a modified version of the John test statistic.}
}
\value{
A list with two values
\describe{
\item{test.statistic}{The test statistic value}
\item{p.value}{The p-value}
}
}
\description{
This function computes the Ledoit-Wolf test statistic for testing the shape of covariance matrix in the one-sample case
}
\examples{
library(mvtnorm)
# Nagao's test for identity
x = rmvnorm(n = 20, mean = numeric(100), sigma = diag(100))
john.nagao.statistic(x, test = "identity")
# John's test for sphericity
y = rmvnorm(n = 10, mean = runif(100, 1, 3), sigma = diag(rgamma(100, 10, 3)))
john.nagao.statistic(y, test = "sphericity")
}
\references{
O. Ledoit and M. Wolf. Some hypothesis tests for the covariance matrix when the dimension is large compared to the sample size. The Annals of Statistics, 30(4):1081–1102, 2002.
}
\seealso{
\code{\link{john.nagao.statistic}}
}
