% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nulldist_onesample.R
\name{nulldist.onesample}
\alias{nulldist.onesample}
\title{One sample null distribution generator}
\usage{
nulldist.onesample(
  n,
  p,
  m = 1,
  nnull = 100,
  nrep = 100,
  mean = NULL,
  parallel = TRUE,
  ncores = NULL
)
}
\arguments{
\item{n}{sample size of your data}

\item{p}{original dimension of your data}

\item{m}{projected dimension of your data (Default value is 1)}

\item{nnull}{number of repetitions of averages of pvalues}

\item{nrep}{number of repetitions for each average of pvalues}

\item{mean}{(Optional) Mean vector for the null distribution}

\item{parallel}{Whether to parallelize or not (default is \code{TRUE})}

\item{ncores}{Number of cores to use (If NULL and parallel is TRUE, use 75\% of all cores). If \code{parallel = FALSE}, this argument will be ignored.}
}
\value{
A list with three elements corresponding to different test statistics
\describe{
\item{LRT}{likelihood ratio test statistic}
\item{LW}{Ledoit-Wolf test statistic}
\item{John}{John-Nagao test statistic}
}
}
\description{
This function generates random samples and evaluates the test statistics for the identity hypothesis and reports a bootstrap random sample of average p-values. The generated values can be used to calculate the critical value for rejecting/accepting the null hypothesis
}
\examples{
\dontrun{
  nulldist.onesample(n = 20, p = 100, m = 5)
}
}
