% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nulldist_twosample.R
\name{nulldist.twosample}
\alias{nulldist.twosample}
\title{Two sample null distribution generator}
\usage{
nulldist.twosample(
  n1,
  n2 = NULL,
  p,
  m = 1,
  nnull = 100,
  nrep = 100,
  mean1 = NULL,
  mean2 = NULL,
  parallel = TRUE,
  ncores = NULL
)
}
\arguments{
\item{n1}{sample size of the data from the first group}

\item{n2}{sample size of the data from the second group. If \code{n2 = NULL} (Default value), then n2 is set equal to n1}

\item{p}{original dimension of your data}

\item{m}{projected dimension of your data (Default value is 1)}

\item{nnull}{number of repetitions of averages of pvalues}

\item{nrep}{number of repetitions for each average of pvalues}

\item{mean1}{(Optional) Mean vector of the first group for the null distribution}

\item{mean2}{(Optional) Mean vector of the second group for the null distribution}

\item{parallel}{Whether to parallelize or not (default is \code{TRUE})}

\item{ncores}{Number of cores to use (If NULL and parallel is TRUE, use 75\% of all cores). If \code{parallel = FALSE}, this argument will be ignored.}
}
\value{
A list with three elements corresponding to different test statistics
\describe{
\item{Box-M}{Box's M test statistic}
\item{Wald}{Wald test statistic}
}
}
\description{
This function generates random samples and evaluates the test statistics for the two-sample covariance matrix test and reports a bootstrap random sample of average p-values. The generated values can be used to calculate the critical value for rejecting/accepting the null hypothesis
}
\examples{
\dontrun{
 nulldist.twosample(n = 20, p = 100, m = 5)
}
}
