% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptt_randproj.R
\name{raptt.randproj}
\alias{raptt.randproj}
\title{RAPTT projection matrix generator}
\usage{
raptt.randproj(nrow, ncol, seed = NULL)
}
\arguments{
\item{nrow}{Number of rows in the random matrix to be generated (dimension of projected space)}

\item{ncol}{Number of colums (dimension of original space)}

\item{seed}{Set the seed to replicate the random matrix generated. Default is \code{NULL} and no seed is used.}
}
\value{
An orthogonal matrix of with \code{nrow} rows and \code{ncol} columns.
}
\description{
Generate random matrices of given dimension which have full row and are orthogonal as defined in Srivastava et al. (2016)
}
\examples{
a = raptt.randproj(nrow = 5, ncol = 10)
}
\references{
Srivastava, R., Li, P. and Ruppert, D. RAPTT: An Exact Two-Sample Test in High Dimensions Using Random Projections. Journal of Computational and Graphical Statistics, 25 (3), 2016.
}
