% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schott_test.R
\name{schott.test}
\alias{schott.test}
\title{Schott's test}
\usage{
schott.test(x, y)
}
\arguments{
\item{x}{data matrix with rows representing samples and columns representing variables}

\item{y}{data matrix of second group with same number of columns as \code{x}}
}
\value{
A list with two values
\describe{
\item{test.statistic}{The test statistic value}
\item{p.value}{The p-value}
}
}
\description{
This function computes the Schott's two sample test statistic for testing equality of covariance matrices as described in Schott (need citation)
}
\examples{
library(mvtnorm)
x = rmvnorm(n = 100, mean = numeric(10), sigma = diag(10))
y = rmvnorm(n = 100, mean = numeric(10), sigma = diag(10))
schott.test(x, y)
}
\references{
J. R. Schott. A test for the equality of covariance matrices when the dimension is large relative to the sample sizes. Computational Statistics and Data Analysis, 51(12):6535–6542, 2007.
}
