% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_statistic.R
\name{wald.statistic}
\alias{wald.statistic}
\title{Two-sample Wald test}
\usage{
wald.statistic(x, y)
}
\arguments{
\item{x}{data matrix of first group with rows representing samples and columns representing variables.}

\item{y}{data matrix of second group.}
}
\value{
A list with two values
\describe{
\item{test.statistic}{The test statistic value}
\item{p.value}{The p-value}
}
}
\description{
Wald-type test statistic for comparing the covariance matrices of two independent groups.
}
\examples{
library(mvtnorm)
x = rmvnorm(n = 100, mean = numeric(10), sigma = diag(10))
y = rmvnorm(n = 100, mean = numeric(10), sigma = diag(10))
wald.statistic(x, y)
}
\references{
J. R. Schott. A test for the equality of covariance matrices when the dimension is large relative to the sample sizes. Computational Statistics and Data Analysis, 51(12):6535–6542, 2007.
}
