% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_li_test.R
\name{wu.li.statistic}
\alias{wu.li.statistic}
\title{Wu-Li test statistic}
\usage{
wu.li.statistic(x, y = NULL, nrand = 1000, test = "identity")
}
\arguments{
\item{x}{data matrix of first group with rows representing samples and columns representing variables.}

\item{y}{data matrix of second group with same number of columns as \code{x}. If not provided (default value is NULL), then the one-sample tests are performed as specified.}

\item{nrand}{number of random projections to be generated for the one-dimension projections.}

\item{test}{Type of test to be performed. When \code{y = NULL}, you can specify "identity" (Default) or "sphericity". If \code{y} is provided as a non-null data matrix, any specified value for \code{test} will be over-written to \code{test = "two.sample"}.}
}
\value{
A list with two values
\describe{
\item{test.statistic}{The test statistic value}
\item{crit.value}{The critical value to accept or reject the null hypothesis based on the \code{test.statistic} returned}
}
}
\description{
This function computes the test statistics for one and two sample hypothesis of covariance matrices as described in Wu and Li (need citation)
}
\examples{
library(mvtnorm)
x = rmvnorm(n = 100, mean = numeric(10), sigma = diag(10))
y = rmvnorm(n = 100, mean = numeric(10), sigma = diag(10))
wu.li.statistic(x, y, nrand = 1e2)
}
\references{
T.-L. Wu and P. Li. Projected tests for high-dimensional covariance matrices. Journal of Statistical Planning and Inference, 207:73 – 85, 2020.
}
\seealso{
\code{gev.critvalue}
}
